requirements.txt
streamlit
langchain>=0.0.284
langchain-community>=0.0.21
langchain-mistralai>=0.1.0
mistralai>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
unstructured[all]>=0.9.2
libmagic>=1.0
python-magic>=0.4.27
python-magic-bin>=0.4.14
langdetect>=1.0.9
